+++ package.json
@@ -1,25 +1,27 @@
 {
-  "name": "line-weather-forecast-bot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/yuya-minamide/weather-forecast-bot.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/yuya-minamide/weather-forecast-bot/issues"
-  },
-  "homepage": "https://github.com/yuya-minamide/weather-forecast-bot#readme",
-  "dependencies": {
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2"
-  }
+	"name": "line-weather-forecast-bot",
+	"version": "1.0.0",
+	"description": "",
+	"main": "server.js",
+	"type": "module",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"start": "node src/server.js"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/yuya-minamide/weather-forecast-bot.git"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/yuya-minamide/weather-forecast-bot/issues"
+	},
+	"homepage": "https://github.com/yuya-minamide/weather-forecast-bot#readme",
+	"dependencies": {
+		"dotenv": "^16.0.3",
+		"express": "^4.18.2",
+		"node-fetch": "^3.3.0"
+	}
 }
